+++ changelog.md
@@ -7,6 +7,7 @@ v0.0.18-alpha
 * Fix bug reseting arcane recipe book data on player death or return from the End
 * Added workaround for translucent textures not rendering under Fabulous graphics settings
 * The Hellish Chain ability now only triggers from attacks that actually do damage
+* Add extra audio-visual effects to active ritual altars
 
 v0.0.17-alpha
 -------------